Move table metadata rendering to separate file
