+++ tools/shell/CMakeLists.txt
@@ -25,6 +25,7 @@ set(SHELL_SOURCES
     shell_renderer.cpp
     shell_highlight.cpp
     shell_progress_bar.cpp
+    shell_render_table_metadata.cpp
     shell_windows.cpp)
 
 option(STATIC_LIBCPP "Statically link CLI to libc++" FALSE)
